Fix sentry Sidekiq bug
client is creating the Model instance within a transaction and pushing a job to Sidekiq.
Sidekiq is trying to execute the job before the transaction has actually committed.
Advice from Sidekiq crew : Use Rails's `after_commit :on => :create` hook or move the job creation outside of the transaction block.
See this link :  https://github.com/mperham/sidekiq/wiki/FAQ#why-am-i-seeing-a-lot-of-cant-find-modelname-with-id12345-errors-with-sidekiq

Since this callbacks is after 2 stats :draft and :change_request when Enrollment are on state :submited, to make it  work and Sidekiq sending this email,
we add to:

- remove module datapass_administrator_notifier
- Add a method to call EnrollmentMailer.with(..).notification_email_unknown_software.deliver_later into the enrollment model
 - Add an If event == 'submit' to Enrollment Controller to call the model's method.