Replace https://app-test.moncomptepro.beta.gouv.fr/users/sign-in with localhost:3000 address
Add some comment to explain why we are log out between each test.
This is a behaviour wanted by cypress team since V10.

A lib has been released to recreate the behaviour to Cypress.Cookies.preserveOnce("_session_id")
but it does not work with Cypress V12.
Hope that Cypress team will add the possibility to stay connected between the test event if it's not recommended by them.