Backend: fix how we archive enrollments without user
`archive!` method bypass the state machine, which is not a good
practice.

However, there's a global `after_transition` which use `user_id` to
affects the owner of the action, but in the case of the worker there is
no owner.

We have to use `try` within this `after_transition` to make it pass.

Tests have been refactored but it tests the same behavior as before.